feat: Reorganize aliases of session subcommands
