+++ src/ai/backend/client/cli/__init__.py
@@ -1,9 +1,9 @@
 from . import main  # noqa
 from . import config  # noqa
-from . import admin  # noqa
 from . import session  # noqa
 from . import session_template  # noqa
 from . import vfolder  # noqa
 from . import dotfile  # noqa
 from . import server_log  # noqa
-from . import app, logs, manager, proxy, run # noqa
+from . import admin  # noqa
+from . import app, logs, proxy, run # noqa
