+++ src/ai/backend/client/cli/run.py
@@ -24,7 +24,6 @@ import tabulate as tabulate_mod
 from tabulate import tabulate
 
 from .main import main
-from .admin.session import info as session_info
 from ..config import local_cache_path
 from ..compat import asyncio_run, current_loop
 from ..exceptions import BackendError, BackendAPIError
@@ -205,7 +204,7 @@ def _noop(*args, **kwargs):
     pass
 
 
-def _format_stats(stats):
+def format_stats(stats):
     formatted = []
     version = stats.pop('version', 1)
     stats.pop('status')
@@ -260,7 +259,7 @@ def _format_stats(stats):
     return tabulate(formatted)
 
 
-def _prepare_resource_arg(resources):
+def prepare_resource_arg(resources):
     if resources:
         resources = {k: v for k, v in map(lambda s: s.split('=', 1), resources)}
     else:
@@ -268,7 +267,7 @@ def _prepare_resource_arg(resources):
     return resources
 
 
-def _prepare_env_arg(env):
+def prepare_env_arg(env):
     if env is not None:
         envs = {k: v for k, v in map(lambda s: s.split('=', 1), env)}
     else:
@@ -276,7 +275,7 @@ def _prepare_env_arg(env):
     return envs
 
 
-def _prepare_mount_arg(
+def prepare_mount_arg(
     mount_args: Optional[Sequence[str]]
 ) -> Tuple[Sequence[str], Mapping[str, str]]:
     """
@@ -423,10 +422,10 @@ def run(image, files, name,                                 # base args
               '"-c" option if run without files.', file=sys.stderr)
         sys.exit(1)
 
-    envs = _prepare_env_arg(env)
-    resources = _prepare_resource_arg(resources)
-    resource_opts = _prepare_resource_arg(resource_opts)
-    mount, mount_map = _prepare_mount_arg(mount)
+    envs = prepare_env_arg(env)
+    resources = prepare_resource_arg(resources)
+    resource_opts = prepare_resource_arg(resource_opts)
+    mount, mount_map = prepare_mount_arg(mount)
 
     if env_range is None: env_range = []      # noqa
     if build_range is None: build_range = []  # noqa
@@ -570,7 +569,7 @@ def run(image, files, name,                                 # base args
                     _stats = ret.get('stats', None) if ret else None
                     if _stats:
                         print('[{0}] Statistics:\n{1}'
-                              .format(idx, _format_stats(_stats)))
+                              .format(idx, format_stats(_stats)))
                     else:
                         print('[{0}] Statistics is not available.'.format(idx))
 
@@ -692,7 +691,7 @@ def run(image, files, name,                                 # base args
                     if stats:
                         _stats = ret.get('stats', None) if ret else None
                         if _stats:
-                            stats_str = _format_stats(_stats)
+                            stats_str = format_stats(_stats)
                             print(format_info('[{0}] Statistics:'.format(idx)) +
                                   '\n{0}'.format(stats_str))
                             if is_multi:
@@ -771,449 +770,3 @@ def run(image, files, name,                                 # base args
             asyncio_run(_run_cases())
     except Exception as e:
         print_fail('{0}'.format(e))
-
-
-@main.command()
-@click.argument('image')
-@click.option('-t', '--name', '--client-token', metavar='NAME',
-              help='Specify a human-readable session name. '
-                   'If not set, a random hex string is used.')
-@click.option('-o', '--owner', '--owner-access-key', metavar='ACCESS_KEY',
-              help='Set the owner of the target session explicitly.')
-# job scheduling options
-@click.option('--type', metavar='SESSTYPE',
-              type=click.Choice(['batch', 'interactive']),
-              default='interactive',
-              help='Either batch or interactive')
-@click.option('--starts-at', metavar='STARTS_AT', type=str, default=None,
-              help='Let session to be started at a specific or relative time.')
-@click.option('-c', '--startup-command', metavar='COMMAND',
-              help='Set the command to execute for batch-type sessions.')
-@click.option('--enqueue-only', is_flag=True,
-              help='Enqueue the session and return immediately without waiting for its startup.')
-@click.option('--max-wait', metavar='SECONDS', type=int, default=0,
-              help='The maximum duration to wait until the session starts.')
-@click.option('--no-reuse', is_flag=True,
-              help='Do not reuse existing sessions but return an error.')
-# execution environment
-@click.option('-e', '--env', metavar='KEY=VAL', type=str, multiple=True,
-              help='Environment variable (may appear multiple times)')
-# extra options
-@click.option('--bootstrap-script', metavar='PATH', type=click.File('r'), default=None,
-              help='A user-defined script to execute on startup.')
-@click.option('--tag', type=str, default=None,
-              help='User-defined tag string to annotate sessions.')
-# resource spec
-@click.option('-v', '--volume', '-m', '--mount', 'mount',
-              metavar='NAME[=PATH]', type=str, multiple=True,
-              help='User-owned virtual folder names to mount. '
-                   'If path is not provided, virtual folder will be mounted under /home/work. '
-                   'All virtual folders can only be mounted under /home/work. ')
-@click.option('--scaling-group', '--sgroup', type=str, default=None,
-              help='The scaling group to execute session. If not specified, '
-                   'all available scaling groups are included in the scheduling.')
-@click.option('-r', '--resources', metavar='KEY=VAL', type=str, multiple=True,
-              help='Set computation resources used by the session '
-                   '(e.g: -r cpu=2 -r mem=256 -r gpu=1).'
-                   '1 slot of cpu/gpu represents 1 core. '
-                   'The unit of mem(ory) is MiB.')
-@click.option('--cluster-size', metavar='NUMBER', type=int, default=1,
-              help='The size of cluster in number of containers.')
-@click.option('--cluster-mode', metavar='MODE',
-              type=click.Choice(['single-node', 'multi-node']), default='single-node',
-              help='The mode of clustering.')
-@click.option('--resource-opts', metavar='KEY=VAL', type=str, multiple=True,
-              help='Resource options for creating compute session '
-                   '(e.g: shmem=64m)')
-# resource grouping
-@click.option('-d', '--domain', metavar='DOMAIN_NAME', default=None,
-              help='Domain name where the session will be spawned. '
-                   'If not specified, config\'s domain name will be used.')
-@click.option('-g', '--group', metavar='GROUP_NAME', default=None,
-              help='Group name where the session is spawned. '
-                   'User should be a member of the group to execute the code.')
-@click.option('--preopen',  default=None,
-              help='Pre-open service ports')
-def start(image, name, owner,                                 # base args
-          type, starts_at, startup_command, enqueue_only, max_wait, no_reuse,  # job scheduling options
-          env,                                            # execution environment
-          bootstrap_script, tag,                          # extra options
-          mount, scaling_group, resources,                # resource spec
-          cluster_size, cluster_mode,
-          resource_opts,
-          domain, group, preopen):                        # resource grouping
-    """
-    Prepare and start a single compute session without executing codes.
-    You may use the created session to execute codes using the "run" command
-    or connect to an application service provided by the session using the "app"
-    command.
-
-
-    \b
-    IMAGE: The name (and version/platform tags appended after a colon) of session
-           runtime or programming language.
-    """
-    if name is None:
-        name = f'pysdk-{secrets.token_hex(5)}'
-    else:
-        name = name
-
-    ######
-    envs = _prepare_env_arg(env)
-    resources = _prepare_resource_arg(resources)
-    resource_opts = _prepare_resource_arg(resource_opts)
-    mount, mount_map = _prepare_mount_arg(mount)
-    preopen_ports = [] if preopen is None else list(map(int, preopen.split(',')))
-    with Session() as session:
-        try:
-            compute_session = session.ComputeSession.get_or_create(
-                image,
-                name=name,
-                type_=type,
-                starts_at=starts_at,
-                enqueue_only=enqueue_only,
-                max_wait=max_wait,
-                no_reuse=no_reuse,
-                cluster_size=cluster_size,
-                cluster_mode=cluster_mode,
-                mounts=mount,
-                mount_map=mount_map,
-                envs=envs,
-                startup_command=startup_command,
-                resources=resources,
-                resource_opts=resource_opts,
-                owner_access_key=owner,
-                domain_name=domain,
-                group_name=group,
-                scaling_group=scaling_group,
-                bootstrap_script=bootstrap_script.read() if bootstrap_script is not None else None,
-                tag=tag,
-                preopen_ports=preopen_ports)
-        except Exception as e:
-            print_error(e)
-            sys.exit(1)
-        else:
-            if compute_session.status == 'PENDING':
-                print_info('Session ID {0} is enqueued for scheduling.'
-                           .format(compute_session.id))
-            elif compute_session.status == 'SCHEDULED':
-                print_info('Session ID {0} is scheduled and about to be started.'
-                           .format(compute_session.id))
-                return
-            elif compute_session.status == 'RUNNING':
-                if compute_session.created:
-                    print_info('Session ID {0} is created and ready.'
-                               .format(compute_session.id))
-                else:
-                    print_info('Session ID {0} is already running and ready.'
-                               .format(compute_session.id))
-                if compute_session.service_ports:
-                    print_info('This session provides the following app services: ' +
-                               ', '.join(sport['name']
-                                         for sport in compute_session.service_ports))
-            elif compute_session.status == 'TERMINATED':
-                print_warn('Session ID {0} is already terminated.\n'
-                           'This may be an error in the compute_session image.'
-                           .format(compute_session.id))
-            elif compute_session.status == 'TIMEOUT':
-                print_info('Session ID {0} is still on the job queue.'
-                           .format(compute_session.id))
-            elif compute_session.status in ('ERROR', 'CANCELLED'):
-                print_fail('Session ID {0} has an error during scheduling/startup or cancelled.'
-                           .format(compute_session.id))
-
-
-@main.command()
-@click.argument('template_id')
-@click.option('-t', '--name', '--client-token', metavar='NAME',
-              default=undefined,
-              help='Specify a human-readable session name. '
-                   'If not set, a random hex string is used.')
-@click.option('-o', '--owner', '--owner-access-key', metavar='ACCESS_KEY',
-              default=undefined,
-              help='Set the owner of the target session explicitly.')
-# job scheduling options
-@click.option('--type', 'type_', metavar='SESSTYPE',
-              type=click.Choice(['batch', 'interactive', undefined]),  # type: ignore
-              default=undefined,
-              help='Either batch or interactive')
-@click.option('--starts_at', metavar='STARTS_AT', type=str, default=None,
-              help='Let session to be started at a specific or relative time.')
-@click.option('-i', '--image', default=undefined,
-              help='Set compute_session image to run.')
-@click.option('-c', '--startup-command', metavar='COMMAND',
-              default=undefined,
-              help='Set the command to execute for batch-type sessions.')
-@click.option('--enqueue-only', is_flag=True,
-              help='Enqueue the session and return immediately without waiting for its startup.')
-@click.option('--max-wait', metavar='SECONDS', type=int,
-              default=-1,
-              help='The maximum duration to wait until the session starts.')
-@click.option('--no-reuse', is_flag=True,
-              help='Do not reuse existing sessions but return an error.')
-# execution environment
-@click.option('-e', '--env', metavar='KEY=VAL', type=str, multiple=True,
-              help='Environment variable (may appear multiple times)')
-# extra options
-@click.option('--tag', type=str,
-              default='$NODEF$',
-              help='User-defined tag string to annotate sessions.')
-# resource spec
-@click.option('-m', '--mount', metavar='NAME[=PATH]', type=str, multiple=True,
-              help='User-owned virtual folder names to mount. '
-                   'If path is not provided, virtual folder will be mounted under /home/work. '
-                   'All virtual folders can only be mounted under /home/work. ')
-@click.option('--scaling-group', '--sgroup', type=str,
-              default='$NODEF$',
-              help='The scaling group to execute session. If not specified, '
-                   'all available scaling groups are included in the scheduling.')
-@click.option('-r', '--resources', metavar='KEY=VAL', type=str, multiple=True,
-              help='Set computation resources used by the session '
-                   '(e.g: -r cpu=2 -r mem=256 -r gpu=1).'
-                   '1 slot of cpu/gpu represents 1 core. '
-                   'The unit of mem(ory) is MiB.')
-@click.option('--cluster-size', metavar='NUMBER', type=int,
-              default=-1,
-              help='The size of cluster in number of containers.')
-@click.option('--resource-opts', metavar='KEY=VAL', type=str, multiple=True,
-              help='Resource options for creating compute session '
-                   '(e.g: shmem=64m)')
-# resource grouping
-@click.option('-d', '--domain', metavar='DOMAIN_NAME', default=None,
-              help='Domain name where the session will be spawned. '
-                   'If not specified, config\'s domain name will be used.')
-@click.option('-g', '--group', metavar='GROUP_NAME', default=None,
-              help='Group name where the session is spawned. '
-                   'User should be a member of the group to execute the code.')
-@click.option('--no-mount', is_flag=True,
-              help='If specified, client.py will tell server not to mount '
-                   'any vFolders specified at template,')
-@click.option('--no-env', is_flag=True,
-              help='If specified, client.py will tell server not to add '
-                   'any environs specified at template,')
-@click.option('--no-resource', is_flag=True,
-              help='If specified, client.py will tell server not to add '
-                   'any resource specified at template,')
-def start_template(
-    template_id, name, owner,        # base args
-    type_, starts_at, image, startup_command, enqueue_only, max_wait, no_reuse,  # job scheduling options
-    env,                                            # execution environment
-    tag,                                            # extra options
-    mount, scaling_group, resources, cluster_size,  # resource spec
-    resource_opts,
-    domain, group,                                  # resource grouping
-    no_mount, no_env, no_resource,
-):
-    """
-    Prepare and start a single compute session without executing codes.
-    You may use the created session to execute codes using the "run" command
-    or connect to an application service provided by the session using the "app"
-    command.
-
-
-    \b
-    IMAGE: The name (and version/platform tags appended after a colon) of session
-           runtime or programming language.
-    """
-    if name is undefined:
-        name = f'pysdk-{secrets.token_hex(5)}'
-    else:
-        name = name
-
-    if max_wait == -1:
-        max_wait = undefined
-    if tag == '$NODEF$':
-        tag = undefined
-    if scaling_group == '$NODEF$':
-        scaling_group = undefined
-    if cluster_size == -1:
-        cluster_size = undefined
-
-    ######
-
-    envs = _prepare_env_arg(env) if len(env) > 0 or no_env else undefined
-    resources = _prepare_resource_arg(resources) if len(resources) > 0 or no_resource else undefined
-    resource_opts = (_prepare_resource_arg(resource_opts)
-                     if len(resource_opts) > 0 or no_resource else undefined)
-    mount, mount_map = (_prepare_mount_arg(mount)
-                        if len(mount) > 0 or no_mount else (undefined, undefined))
-    with Session() as session:
-        try:
-            compute_session = session.ComputeSession.create_from_template(
-                template_id,
-                image=image,
-                name=name,
-                type_=type_,
-                starts_at=starts_at,
-                enqueue_only=enqueue_only,
-                max_wait=max_wait,
-                no_reuse=no_reuse,
-                cluster_size=cluster_size,
-                mounts=mount,
-                mount_map=mount_map,
-                envs=envs,
-                startup_command=startup_command,
-                resources=resources,
-                resource_opts=resource_opts,
-                owner_access_key=owner,
-                domain_name=domain,
-                group_name=group,
-                scaling_group=scaling_group,
-                tag=tag)
-        except Exception as e:
-            print_error(e)
-            sys.exit(1)
-        else:
-            if compute_session.status == 'PENDING':
-                print_info('Session ID {0} is enqueued for scheduling.'
-                           .format(name))
-            elif compute_session.status == 'SCHEDULED':
-                print_info('Session ID {0} is scheduled and about to be started.'
-                           .format(name))
-                return
-            elif compute_session.status == 'RUNNING':
-                if compute_session.created:
-                    print_info('Session ID {0} is created and ready.'
-                               .format(name))
-                else:
-                    print_info('Session ID {0} is already running and ready.'
-                               .format(name))
-                if compute_session.service_ports:
-                    print_info('This session provides the following app services: ' +
-                               ', '.join(sport['name']
-                                         for sport in compute_session.service_ports))
-            elif compute_session.status == 'TERMINATED':
-                print_warn('Session ID {0} is already terminated.\n'
-                           'This may be an error in the compute_session image.'
-                           .format(name))
-            elif compute_session.status == 'TIMEOUT':
-                print_info('Session ID {0} is still on the job queue.'
-                           .format(name))
-            elif compute_session.status in ('ERROR', 'CANCELLED'):
-                print_fail('Session ID {0} has an error during scheduling/startup or cancelled.'
-                           .format(name))
-
-
-@main.command(aliases=['rm', 'kill'])
-@click.argument('session_names', metavar='SESSID', nargs=-1)
-@click.option('-f', '--forced', is_flag=True,
-              help='Force-terminate the errored sessions (only allowed for admins)')
-@click.option('-o', '--owner', '--owner-access-key', metavar='ACCESS_KEY',
-              help='Specify the owner of the target session explicitly.')
-@click.option('-s', '--stats', is_flag=True,
-              help='Show resource usage statistics after termination')
-def terminate(session_names, forced, owner, stats):
-    """
-    Terminate the given session.
-
-    SESSID: session ID given/generated when creating the session.
-    """
-    if len(session_names) == 0:
-        print_warn('Specify at least one session ID. Check usage with "-h" option.')
-        sys.exit(1)
-    print_wait('Terminating the session(s)...')
-    with Session() as session:
-        has_failure = False
-        for session_name in session_names:
-            try:
-                compute_session = session.ComputeSession(session_name, owner)
-                ret = compute_session.destroy(forced=forced)
-            except BackendAPIError as e:
-                print_error(e)
-                if e.status == 404:
-                    print_info(
-                        'If you are an admin, use "-o" / "--owner" option '
-                        'to terminate other user\'s session.')
-                has_failure = True
-            except Exception as e:
-                print_error(e)
-                has_failure = True
-        else:
-            if not has_failure:
-                print_done('Done.')
-            if stats:
-                stats = ret.get('stats', None) if ret else None
-                if stats:
-                    print(_format_stats(stats))
-                else:
-                    print('Statistics is not available.')
-        if has_failure:
-            sys.exit(1)
-
-
-@main.command()
-@click.argument('session_names', metavar='SESSION_NAME', nargs=-1)
-def restart(session_names):
-    """
-    Restart the given session.
-
-    SESSID: session ID given/generated when creating the session.
-    """
-    if len(session_names) == 0:
-        print_warn('Specify at least one session ID. Check usage with "-h" option.')
-        sys.exit(1)
-    print_wait('Restarting the session(s)...')
-    with Session() as session:
-        has_failure = False
-        for session_name in session_names:
-            try:
-                compute_session = session.ComputeSession(session_name)
-                compute_session.restart()
-            except BackendAPIError as e:
-                print_error(e)
-                if e.status == 404:
-                    print_info(
-                        'If you are an admin, use "-o" / "--owner" option '
-                        'to terminate other user\'s session.')
-                has_failure = True
-            except Exception as e:
-                print_error(e)
-                has_failure = True
-        else:
-            if not has_failure:
-                print_done('Done.')
-        if has_failure:
-            sys.exit(1)
-
-
-@main.command()
-@click.argument('session_id', metavar='SESSID')
-@click.pass_context
-def info(ctx, session_id):
-    """
-    Show detailed information for a running compute session.
-    This is an alias of the "admin session <sess_id>" command.
-
-    SESSID: session ID or its alias given when creating the session.
-    """
-    ctx.forward(session_info)
-
-
-@main.command()
-@click.argument('session_name_or_id', metavar='SESSION_ID_OR_NAME')
-@click.option('-o', '--owner', '--owner-access-key', 'owner_access_key', metavar='ACCESS_KEY',
-              help='Specify the owner of the target session explicitly.')
-@click.option('--scope', type=click.Choice(['*', 'session', 'kernel']), default='*',
-              help='Filter the events by kernel-specific ones or session-specific ones.')
-def events(session_name_or_id, owner_access_key, scope):
-    """
-    Monitor the lifecycle events of a compute session.
-
-    SESSID: session ID or its alias given when creating the session.
-    """
-
-    async def _run_events():
-        async with AsyncSession() as session:
-            try:
-                session_id = uuid.UUID(session_name_or_id)
-                compute_session = session.ComputeSession.from_session_id(session_id)
-            except ValueError:
-                compute_session = session.ComputeSession(session_name_or_id, owner_access_key)
-            async with compute_session.listen_events(scope=scope) as response:
-                async for ev in response:
-                    print(click.style(ev.event, fg='cyan', bold=True), json.loads(ev.data))
-
-    try:
-        asyncio_run(_run_events())
-    except Exception as e:
-        print_error(e)
