+++ src/ai/backend/client/cli/session.py
@@ -1,22 +1,475 @@
+from __future__ import annotations
+
 from datetime import datetime
 import json
 from pathlib import Path
+import secrets
 import subprocess
 import sys
+import uuid
 
 import click
 from humanize import naturalsize
 from tabulate import tabulate
 
 from .main import main
-from .pretty import print_wait, print_done, print_error, print_fail
+from .pretty import print_wait, print_done, print_error, print_fail, print_info, print_warn
 from .ssh import container_ssh_ctx
-from ..session import Session
+from .run import format_stats, prepare_env_arg, prepare_resource_arg, prepare_mount_arg
+from ..compat import asyncio_run
+from ..exceptions import BackendAPIError
+from ..session import Session, AsyncSession
+from ..types import undefined
 
 
 @main.group()
 def session():
-    '''Provides compute session operations'''
+    """Set of compute session operations"""
+
+
+def _create_cmd(docs: str = None):
+
+    @click.argument('image')
+    @click.option('-t', '--name', '--client-token', metavar='NAME',
+                  help='Specify a human-readable session name. '
+                       'If not set, a random hex string is used.')
+    @click.option('-o', '--owner', '--owner-access-key', metavar='ACCESS_KEY',
+                  help='Set the owner of the target session explicitly.')
+    # job scheduling options
+    @click.option('--type', metavar='SESSTYPE',
+                  type=click.Choice(['batch', 'interactive']),
+                  default='interactive',
+                  help='Either batch or interactive')
+    @click.option('--starts-at', metavar='STARTS_AT', type=str, default=None,
+                  help='Let session to be started at a specific or relative time.')
+    @click.option('-c', '--startup-command', metavar='COMMAND',
+                  help='Set the command to execute for batch-type sessions.')
+    @click.option('--enqueue-only', is_flag=True,
+                  help='Enqueue the session and return immediately without waiting for its startup.')
+    @click.option('--max-wait', metavar='SECONDS', type=int, default=0,
+                  help='The maximum duration to wait until the session starts.')
+    @click.option('--no-reuse', is_flag=True,
+                  help='Do not reuse existing sessions but return an error.')
+    # execution environment
+    @click.option('-e', '--env', metavar='KEY=VAL', type=str, multiple=True,
+                  help='Environment variable (may appear multiple times)')
+    # extra options
+    @click.option('--bootstrap-script', metavar='PATH', type=click.File('r'), default=None,
+                  help='A user-defined script to execute on startup.')
+    @click.option('--tag', type=str, default=None,
+                  help='User-defined tag string to annotate sessions.')
+    # resource spec
+    @click.option('-v', '--volume', '-m', '--mount', 'mount',
+                  metavar='NAME[=PATH]', type=str, multiple=True,
+                  help='User-owned virtual folder names to mount. '
+                       'If path is not provided, virtual folder will be mounted under /home/work. '
+                       'All virtual folders can only be mounted under /home/work. ')
+    @click.option('--scaling-group', '--sgroup', type=str, default=None,
+                  help='The scaling group to execute session. If not specified, '
+                       'all available scaling groups are included in the scheduling.')
+    @click.option('-r', '--resources', metavar='KEY=VAL', type=str, multiple=True,
+                  help='Set computation resources used by the session '
+                       '(e.g: -r cpu=2 -r mem=256 -r gpu=1).'
+                       '1 slot of cpu/gpu represents 1 core. '
+                       'The unit of mem(ory) is MiB.')
+    @click.option('--cluster-size', metavar='NUMBER', type=int, default=1,
+                  help='The size of cluster in number of containers.')
+    @click.option('--cluster-mode', metavar='MODE',
+                  type=click.Choice(['single-node', 'multi-node']), default='single-node',
+                  help='The mode of clustering.')
+    @click.option('--resource-opts', metavar='KEY=VAL', type=str, multiple=True,
+                  help='Resource options for creating compute session '
+                       '(e.g: shmem=64m)')
+    # resource grouping
+    @click.option('-d', '--domain', metavar='DOMAIN_NAME', default=None,
+                  help='Domain name where the session will be spawned. '
+                       'If not specified, config\'s domain name will be used.')
+    @click.option('-g', '--group', metavar='GROUP_NAME', default=None,
+                  help='Group name where the session is spawned. '
+                       'User should be a member of the group to execute the code.')
+    @click.option('--preopen',  default=None,
+                  help='Pre-open service ports')
+    def create(
+        image, name, owner,                                 # base args
+        type, starts_at, startup_command, enqueue_only, max_wait, no_reuse,  # job scheduling options
+        env,                                            # execution environment
+        bootstrap_script, tag,                          # extra options
+        mount, scaling_group, resources,                # resource spec
+        cluster_size, cluster_mode,
+        resource_opts,
+        domain, group, preopen,                         # resource grouping
+    ) -> None:
+        """
+        Prepare and start a single compute session without executing codes.
+        You may use the created session to execute codes using the "run" command
+        or connect to an application service provided by the session using the "app"
+        command.
+
+
+        \b
+        IMAGE: The name (and version/platform tags appended after a colon) of session
+               runtime or programming language.
+        """
+        if name is None:
+            name = f'pysdk-{secrets.token_hex(5)}'
+        else:
+            name = name
+
+        ######
+        envs = prepare_env_arg(env)
+        resources = prepare_resource_arg(resources)
+        resource_opts = prepare_resource_arg(resource_opts)
+        mount, mount_map = prepare_mount_arg(mount)
+        preopen_ports = [] if preopen is None else list(map(int, preopen.split(',')))
+        with Session() as session:
+            try:
+                compute_session = session.ComputeSession.get_or_create(
+                    image,
+                    name=name,
+                    type_=type,
+                    starts_at=starts_at,
+                    enqueue_only=enqueue_only,
+                    max_wait=max_wait,
+                    no_reuse=no_reuse,
+                    cluster_size=cluster_size,
+                    cluster_mode=cluster_mode,
+                    mounts=mount,
+                    mount_map=mount_map,
+                    envs=envs,
+                    startup_command=startup_command,
+                    resources=resources,
+                    resource_opts=resource_opts,
+                    owner_access_key=owner,
+                    domain_name=domain,
+                    group_name=group,
+                    scaling_group=scaling_group,
+                    bootstrap_script=bootstrap_script.read() if bootstrap_script is not None else None,
+                    tag=tag,
+                    preopen_ports=preopen_ports,
+                )
+            except Exception as e:
+                print_error(e)
+                sys.exit(1)
+            else:
+                if compute_session.status == 'PENDING':
+                    print_info('Session ID {0} is enqueued for scheduling.'
+                               .format(compute_session.id))
+                elif compute_session.status == 'SCHEDULED':
+                    print_info('Session ID {0} is scheduled and about to be started.'
+                               .format(compute_session.id))
+                    return
+                elif compute_session.status == 'RUNNING':
+                    if compute_session.created:
+                        print_info('Session ID {0} is created and ready.'
+                                   .format(compute_session.id))
+                    else:
+                        print_info('Session ID {0} is already running and ready.'
+                                   .format(compute_session.id))
+                    if compute_session.service_ports:
+                        print_info('This session provides the following app services: ' +
+                                   ', '.join(sport['name']
+                                             for sport in compute_session.service_ports))
+                elif compute_session.status == 'TERMINATED':
+                    print_warn('Session ID {0} is already terminated.\n'
+                               'This may be an error in the compute_session image.'
+                               .format(compute_session.id))
+                elif compute_session.status == 'TIMEOUT':
+                    print_info('Session ID {0} is still on the job queue.'
+                               .format(compute_session.id))
+                elif compute_session.status in ('ERROR', 'CANCELLED'):
+                    print_fail('Session ID {0} has an error during scheduling/startup or cancelled.'
+                               .format(compute_session.id))
+
+    if docs is not None:
+        create.__doc__ = docs
+    return create
+
+
+main.command(aliases=['start'])(_create_cmd(docs="Alias of \"session create\""))
+session.command()(_create_cmd())
+
+
+def _create_from_template_cmd(docs: str = None):
+
+    @click.argument('template_id')
+    @click.option('-t', '--name', '--client-token', metavar='NAME',
+                  default=undefined,
+                  help='Specify a human-readable session name. '
+                       'If not set, a random hex string is used.')
+    @click.option('-o', '--owner', '--owner-access-key', metavar='ACCESS_KEY',
+                  default=undefined,
+                  help='Set the owner of the target session explicitly.')
+    # job scheduling options
+    @click.option('--type', 'type_', metavar='SESSTYPE',
+                  type=click.Choice(['batch', 'interactive', undefined]),  # type: ignore
+                  default=undefined,
+                  help='Either batch or interactive')
+    @click.option('--starts_at', metavar='STARTS_AT', type=str, default=None,
+                  help='Let session to be started at a specific or relative time.')
+    @click.option('-i', '--image', default=undefined,
+                  help='Set compute_session image to run.')
+    @click.option('-c', '--startup-command', metavar='COMMAND',
+                  default=undefined,
+                  help='Set the command to execute for batch-type sessions.')
+    @click.option('--enqueue-only', is_flag=True,
+                  help='Enqueue the session and return immediately without waiting for its startup.')
+    @click.option('--max-wait', metavar='SECONDS', type=int,
+                  default=-1,
+                  help='The maximum duration to wait until the session starts.')
+    @click.option('--no-reuse', is_flag=True,
+                  help='Do not reuse existing sessions but return an error.')
+    # execution environment
+    @click.option('-e', '--env', metavar='KEY=VAL', type=str, multiple=True,
+                  help='Environment variable (may appear multiple times)')
+    # extra options
+    @click.option('--tag', type=str,
+                  default='$NODEF$',
+                  help='User-defined tag string to annotate sessions.')
+    # resource spec
+    @click.option('-m', '--mount', metavar='NAME[=PATH]', type=str, multiple=True,
+                  help='User-owned virtual folder names to mount. '
+                       'If path is not provided, virtual folder will be mounted under /home/work. '
+                       'All virtual folders can only be mounted under /home/work. ')
+    @click.option('--scaling-group', '--sgroup', type=str,
+                  default='$NODEF$',
+                  help='The scaling group to execute session. If not specified, '
+                       'all available scaling groups are included in the scheduling.')
+    @click.option('-r', '--resources', metavar='KEY=VAL', type=str, multiple=True,
+                  help='Set computation resources used by the session '
+                       '(e.g: -r cpu=2 -r mem=256 -r gpu=1).'
+                       '1 slot of cpu/gpu represents 1 core. '
+                       'The unit of mem(ory) is MiB.')
+    @click.option('--cluster-size', metavar='NUMBER', type=int,
+                  default=-1,
+                  help='The size of cluster in number of containers.')
+    @click.option('--resource-opts', metavar='KEY=VAL', type=str, multiple=True,
+                  help='Resource options for creating compute session '
+                       '(e.g: shmem=64m)')
+    # resource grouping
+    @click.option('-d', '--domain', metavar='DOMAIN_NAME', default=None,
+                  help='Domain name where the session will be spawned. '
+                       'If not specified, config\'s domain name will be used.')
+    @click.option('-g', '--group', metavar='GROUP_NAME', default=None,
+                  help='Group name where the session is spawned. '
+                       'User should be a member of the group to execute the code.')
+    @click.option('--no-mount', is_flag=True,
+                  help='If specified, client.py will tell server not to mount '
+                       'any vFolders specified at template,')
+    @click.option('--no-env', is_flag=True,
+                  help='If specified, client.py will tell server not to add '
+                       'any environs specified at template,')
+    @click.option('--no-resource', is_flag=True,
+                  help='If specified, client.py will tell server not to add '
+                       'any resource specified at template,')
+    def create_from_template(
+        template_id, name, owner,        # base args
+        type_, starts_at, image, startup_command, enqueue_only, max_wait, no_reuse,  # job scheduling options
+        env,                                            # execution environment
+        tag,                                            # extra options
+        mount, scaling_group, resources, cluster_size,  # resource spec
+        resource_opts,
+        domain, group,                                  # resource grouping
+        no_mount, no_env, no_resource,
+    ) -> None:
+        """
+        Prepare and start a single compute session without executing codes.
+        You may use the created session to execute codes using the "run" command
+        or connect to an application service provided by the session using the "app"
+        command.
+
+
+        \b
+        IMAGE: The name (and version/platform tags appended after a colon) of session
+               runtime or programming language.
+        """
+        if name is undefined:
+            name = f'pysdk-{secrets.token_hex(5)}'
+        else:
+            name = name
+
+        if max_wait == -1:
+            max_wait = undefined
+        if tag == '$NODEF$':
+            tag = undefined
+        if scaling_group == '$NODEF$':
+            scaling_group = undefined
+        if cluster_size == -1:
+            cluster_size = undefined
+
+        ######
+
+        envs = prepare_env_arg(env) if len(env) > 0 or no_env else undefined
+        resources = prepare_resource_arg(resources) if len(resources) > 0 or no_resource else undefined
+        resource_opts = (prepare_resource_arg(resource_opts)
+                         if len(resource_opts) > 0 or no_resource else undefined)
+        mount, mount_map = (prepare_mount_arg(mount)
+                            if len(mount) > 0 or no_mount else (undefined, undefined))
+        with Session() as session:
+            try:
+                compute_session = session.ComputeSession.create_from_template(
+                    template_id,
+                    image=image,
+                    name=name,
+                    type_=type_,
+                    starts_at=starts_at,
+                    enqueue_only=enqueue_only,
+                    max_wait=max_wait,
+                    no_reuse=no_reuse,
+                    cluster_size=cluster_size,
+                    mounts=mount,
+                    mount_map=mount_map,
+                    envs=envs,
+                    startup_command=startup_command,
+                    resources=resources,
+                    resource_opts=resource_opts,
+                    owner_access_key=owner,
+                    domain_name=domain,
+                    group_name=group,
+                    scaling_group=scaling_group,
+                    tag=tag)
+            except Exception as e:
+                print_error(e)
+                sys.exit(1)
+            else:
+                if compute_session.status == 'PENDING':
+                    print_info('Session ID {0} is enqueued for scheduling.'
+                               .format(name))
+                elif compute_session.status == 'SCHEDULED':
+                    print_info('Session ID {0} is scheduled and about to be started.'
+                               .format(name))
+                    return
+                elif compute_session.status == 'RUNNING':
+                    if compute_session.created:
+                        print_info('Session ID {0} is created and ready.'
+                                   .format(name))
+                    else:
+                        print_info('Session ID {0} is already running and ready.'
+                                   .format(name))
+                    if compute_session.service_ports:
+                        print_info('This session provides the following app services: ' +
+                                   ', '.join(sport['name']
+                                             for sport in compute_session.service_ports))
+                elif compute_session.status == 'TERMINATED':
+                    print_warn('Session ID {0} is already terminated.\n'
+                               'This may be an error in the compute_session image.'
+                               .format(name))
+                elif compute_session.status == 'TIMEOUT':
+                    print_info('Session ID {0} is still on the job queue.'
+                               .format(name))
+                elif compute_session.status in ('ERROR', 'CANCELLED'):
+                    print_fail('Session ID {0} has an error during scheduling/startup or cancelled.'
+                               .format(name))
+
+    if docs is not None:
+        create_from_template.__doc__ = docs
+    return create_from_template
+
+
+main.command(aliases=['start-from-template'])(
+    _create_from_template_cmd(docs="Alias of \"session create-from-template\"")
+)
+session.command()(_create_from_template_cmd())
+
+
+def _destroy_cmd(docs: str = None):
+
+    @click.argument('session_names', metavar='SESSID', nargs=-1)
+    @click.option('-f', '--forced', is_flag=True,
+                  help='Force-terminate the errored sessions (only allowed for admins)')
+    @click.option('-o', '--owner', '--owner-access-key', metavar='ACCESS_KEY',
+                  help='Specify the owner of the target session explicitly.')
+    @click.option('-s', '--stats', is_flag=True,
+                  help='Show resource usage statistics after termination')
+    def destroy(session_names, forced, owner, stats):
+        """
+        Terminate and destroy the given session.
+
+        SESSID: session ID given/generated when creating the session.
+        """
+        if len(session_names) == 0:
+            print_warn('Specify at least one session ID. Check usage with "-h" option.')
+            sys.exit(1)
+        print_wait('Terminating the session(s)...')
+        with Session() as session:
+            has_failure = False
+            for session_name in session_names:
+                try:
+                    compute_session = session.ComputeSession(session_name, owner)
+                    ret = compute_session.destroy(forced=forced)
+                except BackendAPIError as e:
+                    print_error(e)
+                    if e.status == 404:
+                        print_info(
+                            'If you are an admin, use "-o" / "--owner" option '
+                            'to terminate other user\'s session.')
+                    has_failure = True
+                except Exception as e:
+                    print_error(e)
+                    has_failure = True
+            else:
+                if not has_failure:
+                    print_done('Done.')
+                if stats:
+                    stats = ret.get('stats', None) if ret else None
+                    if stats:
+                        print(format_stats(stats))
+                    else:
+                        print('Statistics is not available.')
+            if has_failure:
+                sys.exit(1)
+
+    if docs is not None:
+        destroy.__doc__ = docs
+    return destroy
+
+
+main.command(aliases=['rm', 'kill'])(_destroy_cmd(docs="Alias of \"session destroy\""))
+session.command(aliases=['rm', 'kill'])(_destroy_cmd())
+
+
+def _restart_cmd(docs: str = None):
+
+    @click.argument('session_refs', metavar='SESSION_REFS', nargs=-1)
+    def restart(session_refs):
+        """
+        Restart the compute session.
+
+        \b
+        SESSION_REF: session ID or name
+        """
+        if len(session_refs) == 0:
+            print_warn('Specify at least one session ID. Check usage with "-h" option.')
+            sys.exit(1)
+        print_wait('Restarting the session(s)...')
+        with Session() as session:
+            has_failure = False
+            for session_ref in session_refs:
+                try:
+                    compute_session = session.ComputeSession(session_ref)
+                    compute_session.restart()
+                except BackendAPIError as e:
+                    print_error(e)
+                    if e.status == 404:
+                        print_info(
+                            'If you are an admin, use "-o" / "--owner" option '
+                            'to terminate other user\'s session.')
+                    has_failure = True
+                except Exception as e:
+                    print_error(e)
+                    has_failure = True
+            else:
+                if not has_failure:
+                    print_done('Done.')
+            if has_failure:
+                sys.exit(1)
+
+    if docs is not None:
+        restart.__doc__ = docs
+    return restart
+
+
+main.command()(_restart_cmd(docs="Alias of \"session restart\""))
+session.command()(_restart_cmd())
 
 
 @session.command()
@@ -24,7 +477,7 @@ def session():
 @click.argument('files', type=click.Path(exists=True), nargs=-1)
 def upload(session_id, files):
     """
-    Upload the given files to the target compute session's home directory.
+    Upload the files to a compute session's home directory.
     If the target directory is in a storage folder mount, the operation is
     effectively same to uploading files to the storage folder.
     It is recommended to use storage folder commands for large file transfers
@@ -56,7 +509,7 @@ def upload(session_id, files):
               help='Destination path to store downloaded file(s)')
 def download(session_id, files, dest):
     """
-    Download files from a running compute session.
+    Download files from a compute session's home directory.
     If the source path is in a storage folder mount, the operation is
     effectively same to downloading files from the storage folder.
     It is recommended to use storage folder commands for large file transfers
@@ -150,7 +603,7 @@ def _ssh_cmd(docs: str = None):
                   help="the port number for localhost")
     @click.pass_context
     def ssh(ctx: click.Context, session_ref: str, port: int) -> None:
-        """Execute the ssh command against the target compute session
+        """Execute the ssh command against the target compute session.
 
         \b
         SESSION_REF: The user-provided name or the unique ID of a running compute session.
@@ -192,7 +645,7 @@ _ssh_cmd_context_settings = {
 main.command(
     context_settings=_ssh_cmd_context_settings,
 )(_ssh_cmd(docs="Alias of \"session ssh\""))
-ssh = session.command(
+session.command(
     context_settings=_ssh_cmd_context_settings,
 )(_ssh_cmd())
 
@@ -215,7 +668,8 @@ def _scp_cmd(docs: str = None):
         port: int,
         recursive: bool,
     ) -> None:
-        """Execute the scp command against the target compute session
+        """
+        Execute the scp command against the target compute session.
 
         \b
         The SRC and DST have the same format with the original scp command,
@@ -270,6 +724,48 @@ def _scp_cmd(docs: str = None):
 main.command(
     context_settings=_ssh_cmd_context_settings,
 )(_scp_cmd(docs="Alias of \"session scp\""))
-scp = session.command(
+session.command(
     context_settings=_ssh_cmd_context_settings,
 )(_scp_cmd())
+
+
+def _events_cmd(docs: str = None):
+
+    @click.argument('session_name_or_id', metavar='SESSION_ID_OR_NAME')
+    @click.option('-o', '--owner', '--owner-access-key', 'owner_access_key', metavar='ACCESS_KEY',
+                  help='Specify the owner of the target session explicitly.')
+    @click.option('--scope', type=click.Choice(['*', 'session', 'kernel']), default='*',
+                  help='Filter the events by kernel-specific ones or session-specific ones.')
+    def events(session_name_or_id, owner_access_key, scope):
+        """
+        Monitor the lifecycle events of a compute session.
+
+        SESSID: session ID or its alias given when creating the session.
+        """
+
+        async def _run_events():
+            async with AsyncSession() as session:
+                try:
+                    session_id = uuid.UUID(session_name_or_id)
+                    compute_session = session.ComputeSession.from_session_id(session_id)
+                except ValueError:
+                    compute_session = session.ComputeSession(session_name_or_id, owner_access_key)
+                async with compute_session.listen_events(scope=scope) as response:
+                    async for ev in response:
+                        print(click.style(ev.event, fg='cyan', bold=True), json.loads(ev.data))
+
+        try:
+            asyncio_run(_run_events())
+        except Exception as e:
+            print_error(e)
+
+    if docs is not None:
+        events.__doc__ = docs
+    return events
+
+
+# Make it available as:
+# - backend.ai events
+# - backend.ai session events
+main.command()(_events_cmd(docs="Alias of \"session events\""))
+session.command()(_events_cmd())
