+++ src/ai/backend/client/cli/ssh.py
@@ -17,7 +17,7 @@ def container_ssh_ctx(session_ref: str, port: int) -> Iterator[Path]:
     key_path = Path(f"~/.ssh/id_{random_id}").expanduser()
     try:
         subprocess.run(
-            ["backend.ai", "download", session_ref, key_filename],
+            ["backend.ai", "session", "download", session_ref, key_filename],
             shell=False,
             check=True,
             stdout=subprocess.PIPE,
@@ -33,7 +33,7 @@ def container_ssh_ctx(session_ref: str, port: int) -> Iterator[Path]:
         # proxy_proc is a background process
         proxy_proc = subprocess.Popen(
             [
-                "backend.ai", "app", session_ref,
+                "backend.ai", "session", "app", session_ref,
                 "sshd", "-b", f"127.0.0.1:{port}",
             ],
             stdout=subprocess.PIPE,
